Switch tests from mocha to tape
