+++ package.json
@@ -20,7 +20,9 @@
     "node": ">=0.10.3 <0.11"
   },
   "scripts": {
-    "test": "./node_modules/.bin/mocha -R spec test/"
+    "mocha": "./node_modules/.bin/mocha -R spec test/",
+    "test": "node test/run.js | tap-spec",
+    "coverage": "node_modules/.bin/istanbul cover test/run.js"
   },
   "dependencies": {
     "JSONStream": "^0.10.0",
@@ -35,10 +37,13 @@
     "wellknown": "^0.3.1"
   },
   "devDependencies": {
+    "istanbul": "^0.3.13",
     "mocha": "^2.1.0",
     "precommit-hook": "1.0.7",
     "proxyquire": "^1.4.0",
-    "should": "^5.0.0"
+    "should": "^5.0.0",
+    "tap-spec": "^3.0.0",
+    "tape": "^4.0.0"
   },
   "repository": {
     "type": "git",
